Overhaul on experience. Backend of levels and ranking.
Users can now level up through experience. Users can now get ranls through levels.

At the moment, experience and levels are global, but ranks are guild-specific. If the bot is to be on multiple servers, more care is to be taken into how the bot handles ranks.

Consider having global and guild experience for users.