fix: Use display_avatar instead of avatar
If a user has not uploaded an avatar, it avatar is None.
display_avatar also matches the avatar a user has on their server
profile, so it may be preferable here.

If need to bypass to get their "true" avatar, use avatar instead of
course, but you must handle the None case described above.