refactor: Prepare for frog leaderboard utilization
Problem: I am trying to work with a conceptualized leaderboard
object without actually having explicitly defined a leaderboard
class with explicit functions. Refactoring all this code is not
productive, and it would be better to do this right, by creating
a leaderboard class object and delegating functionality to said
class.

Current code is messy as fu, and while refactoring has reduced
the redundancy, it's still a giant cluster fu.

Frog leaderboards will probably be delayed while planning for a
leaderboards class is functional, and then re-integrated into
everything else.