feat: add initial inktober submission
in the database, under the table `guild` exists a `inktober_cid` column
that corresponds to a guild id. when set, inktober.py will watch that
channel for inktober submissions whose message contents contain a
specific string `inktober day ##`.

if so, a thumbs up reaction will be added to that message.

the scrapper will look for any messages match the above criteria and will
"do stuff" to it, currently TODO.